Add non TTY stream support
When running in a non TTY environment (for example on AWS while writing to CloudWatch) the progressbar will just print with a newline at the end to show the progress. The rate of printing is reduced to once every second so it doesn’t spam the logs.

This commit also contains a fix so that the progressbar will not crash when using interrupt or clear on finished.